Style compomenent in login jsx
