--- src/components/pages/login/LoginPage.scss
@@ -1,8 +1,0 @@
-.login-form {
-  width: 70%;
-  margin: 0 auto;
-}
-
-.password-forget-link {
-  margin-top: 32px !important;
-}
